requirements.txt

python-telegram-bot==13.15
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.25.0
mplfinance>=0.12.10b0
matplotlib>=3.7.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
requests>=2.31.0
